Move MemoryUtils to its own file
